move distr tests back to module but under a subdir
