+++ modules/distribution/tests/allocation_test.go
@@ -1,4 +1,4 @@
-package distr
+package tests
 
 import (
 	"testing"
