+++ modules/distribution/tests/keeper_test.go
@@ -1,4 +1,4 @@
-package distr
+package tests
 
 import (
 	"testing"
